@dannyfpv 2019-05-20
tested on pixhawk 4 v5
indoor test
Modes Tested

Takeoff: Good
Stabilized Mode: Good.

Pre-flight Procedure:

Flight Test:
Arm and Take off in the stabilized mode no issue
Throttle Test no issue
Pitch, Roll, Yaw no issue
Land no issue

Notes: Armed and took off in stabilized mode


https://review.px4.io/plot_app?log=aa6502c1-4f65-4622-a0f8-562795936e10