@jorge789 2019-05-20
###  Indoor flight test on PixRacer V4

Modes Tested

Takeoff: Good
Stabilized Mode: Good.

Pre-flight Procedure:

Flight Test: 
Arm and Take off in the stabilized mode ✓
Throttle Test ✓
Pitch, Roll, Yaw ✓
Land ✓

### Notes:

Took off in stabilized mode for approximately 5 min, No issues noted good flight in general. The vehicle behaved as expected. 

Flight log: 1
https://review.px4.io/plot_app?log=3d557ab4-1ca7-4bee-a095-8f72371533cf

Flight log: 2
https://review.px4.io/plot_app?log=637bfdf6-7aa9-4ef5-bf97-f9f1ffca0d52


###  Indoor flight test on Pixhawk 2 Cube V3

Modes Tested

Takeoff: Good
Stabilized Mode: Good.

Pre-flight Procedure:

Flight Test: 
Arm and Take off in the stabilized mode ✓
Throttle Test ✓
Pitch, Roll, Yaw ✓
Land ✓

### Notes:

Took off in stabilized mode for approximately 5 min, No issues noted good flight in general. The vehicle behaved as expected. 

Flight log: 1 
https://review.px4.io/plot_app?log=25566deb-76b0-460b-b6b1-5b6109f7d975

Flight log : 2
https://review.px4.io/plot_app?log=19995186-fdae-4faa-b52f-7bc7ab64cffb