@dagar 2019-05-26
The main motivation for this change is saving memory (see https://github.com/PX4/Firmware/pull/11176#issue-243209888 for details), but the performance improvements in uORB are also going to be worthwhile overall.

Here's a comparison of the common check and copy operations.

### C API (orb_check(), orb_copy(), etc)
	orb_check vehicle_local_position: 1000 events, 2746us elapsed, 2.75us avg, min 2us max 4us 0.438us rms
	orb_stat vehicle_local_position: 1000 events, 2319us elapsed, 2.32us avg, min 2us max 3us 0.466us rms
	orb_copy vehicle_local_position: 1000 events, 3928us elapsed, 3.93us avg, min 3us max 5us 0.324us rms

### C++ API
	uORB::Subscription orb_check vehicle_local_position: 1000 events, 907us elapsed, 0.91us avg, min 1us max 2us 0.301us rms
	uORB::Subscription orb_stat vehicle_local_position: 1000 events, 1092us elapsed, 1.09us avg, min 1us max 2us 0.289us rms
	uORB::Subscription orb_copy vehicle_local_position: 1000 events, 2530us elapsed, 2.53us avg, min 2us max 3us 0.499us rms


On a typical setup these operations are called tens of thousands of times every second, so shaving off microseconds from each orb check and copy will actually be a considerable cpu usage improvement (in aggregate).