@LorenzMeier 2019-05-26
Awesome! Just for the sake of curiosity: Coukd you share the output of top for one representative setup?